change skip_bytes size type from int64_t to uint64_t (#854)

+++ tsMuxer/ioContextDemuxer.cpp
@@ -118,17 +118,17 @@ unsigned IOContextDemuxer::get_buffer(uint8_t* binary, unsigned size)
     return static_cast<uint32_t>(dst - binary);
 }
 
-void IOContextDemuxer::skip_bytes(const int64_t size)
+void IOContextDemuxer::skip_bytes(const uint64_t size)
 {
     if (size == 0)
         return;
     uint32_t readedBytes = 0;
     int readRez = 0;
-    int64_t skipLeft = size;
+    uint64_t skipLeft = size;
 
     if (m_curPos < m_bufEnd)
     {
-        const int64_t copyLen = min(m_bufEnd - m_curPos, skipLeft);
+        const uint64_t copyLen = min((uint64_t)(m_bufEnd - m_curPos), skipLeft);
         skipLeft -= copyLen;
         m_curPos += copyLen;
         m_processedBytes += copyLen;
@@ -150,7 +150,7 @@ void IOContextDemuxer::skip_bytes(const int64_t size)
         m_bufEnd = m_curPos + readedBytes;
         if (readedBytes == 0)
             break;
-        const int64_t copyLen = min(readedBytes, skipLeft);
+        const uint64_t copyLen = min((uint64_t)readedBytes, skipLeft);
         m_curPos += copyLen;
         m_processedBytes += copyLen;
         skipLeft -= copyLen;
